[refactor] Move definitions out of lib/attic.rb
NoSingletonError -> lib/attic/error.rb
InstanceMethods  -> lib/attic/instance_methods.rb
ClassMethods     -> lib/attic/class_methods.rb
